+++ ectemplate-servicio-clientes/src/main/resources/application.properties
@@ -1,4 +1,5 @@
 spring.application.name=servicio-clientes
-server.port=8002
+server.port=${PORT:0}
 
+eureka.instance.instance-id=${spring.application.name}:${spring.application.instance_id:${random.value}}
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
