+++ ectemplate-servicio-empleados/pom.xml
@@ -5,10 +5,10 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.3</version>
+		<version>2.5.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>net.ostemplate.app.eureka</groupId>
+	<groupId>net.ectemplate.app.eureka</groupId>
 	<artifactId>ectemplate-servicio-clientes-1</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>ectemplate-servicio-clientes-1</name>
