+++ ectemplate-servicio-empleados/src/main/java/net/ectemplate/app/empleados/OstemplateServicioEmpleadosApplication.java
@@ -1,0 +1,15 @@
+package net.ectemplate.app.empleados;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@EnableEurekaClient
+@SpringBootApplication
+public class OstemplateServicioEmpleadosApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OstemplateServicioEmpleadosApplication.class, args);
+	}
+
+}
