--- ectemplate-servicio-empleados/src/main/java/net/ostemplate/app/eureka/OstemplateServicioClientes1Application.java
@@ -1,13 +1,0 @@
-package net.ostemplate.app.eureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OstemplateServicioClientes1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OstemplateServicioClientes1Application.class, args);
-	}
-
-}
