+++ ectemplate-servicio-empleados/src/main/resources/application.properties
@@ -1,1 +1,7 @@
+spring.application.name=servicio-empleados
+server.port=${PORT:0}
+
+eureka.instance.instance-id=${spring.application.name}:${spring.application.instance_id:${random.value}}
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+
