+++ ectemplate-servicio-gateway/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+spring.application.name=servicio-getway
+server.port=8090
+eureka.instance.instance-id=${spring.application.name}:${spring.application.instance_id:${random.value}}
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+
+
+spring.cloud.gateway.routes[0].id=servicio-productos
+spring.cloud.gateway.routes[0].uri=lb://servicio-productos
+spring.cloud.gateway.routes[0].predicates=Path=/api/productos/**
+spring.cloud.gateway.routes[0].filters=StripPrefix=2
