+++ ectemplate-servicio-gateway/src/test/java/net/ectemplate/app/getway/EctemplateServicioGetwayApplicationTests.java
@@ -1,0 +1,13 @@
+package net.ectemplate.app.getway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EctemplateServicioGetwayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
