--- ectemplate-servicio-persona/src/main/java/net/ectemplate/app/persona/EctemplateServicioPersonaApplication.java
@@ -1,13 +1,0 @@
-package net.ectemplate.app.persona;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EctemplateServicioPersonaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EctemplateServicioPersonaApplication.class, args);
-	}
-
-}
