--- ectemplate-servicio-persona/src/test/java/net/ectemplate/app/persona/EctemplateServicioPersonaApplicationTests.java
@@ -1,13 +1,0 @@
-package net.ectemplate.app.persona;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EctemplateServicioPersonaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
