+++ ectemplate-servicio-personas/src/main/java/net/ectemplate/app/personas/controllers/PersonasController.java
@@ -1,16 +1,14 @@
-package net.ostemplate.app.clientes.controllers;
+package net.ectemplate.app.personas.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController 
-public class ClientesController {
+@RestController
+public class PersonasController {
 	
 	@GetMapping("/prueba")
 	public String prueba() {
-		
 		return "OK";
 	}
-	
 
 }
