+++ ectemplate-servicio-personas/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=servicio-productos
+spring.application.name=servicio-personas
 server.port=${PORT:0}
 eureka.instance.instance-id=${spring.application.name}:${spring.application.instance_id:${random.value}}
-eureka.client.service-url.defautlZone=http://localhost:8761/eureka
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka
