+++ ectemplate-servicio-personas/src/test/java/net/ectemplate/app/personas/OstemplateServicioPersonasApplicationTests.java
@@ -1,0 +1,13 @@
+package net.ectemplate.app.personas;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OstemplateServicioPersonasApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
