+++ ectemplate-servicio-productos/pom.xml
@@ -5,10 +5,10 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.5.2</version>
+		<version>2.5.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>net.ostemplate.app.productos</groupId>
+	<groupId>net.ectemplate.app.productos</groupId>
 	<artifactId>ectemplate-servicio-productos</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>ectemplate-servicio-productos</name>
